pyproject.toml
[project]
name = "macopenworker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp",
    "typer",
    "rich",
    "python-docx",
    "openpyxl",
    "pypdf",
    "chromadb",
    "sentence-transformers",
    "python-dotenv",
    "openai",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.8.0",
]
